+++ pkg/oc/cli/cli.go
@@ -19,6 +19,15 @@ import (
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	ktemplates "k8s.io/kubernetes/pkg/kubectl/util/templates"
 
+	"github.com/openshift/oc/pkg/cli/experimental/dockergc"
+	"github.com/openshift/oc/pkg/cli/expose"
+	"github.com/openshift/oc/pkg/cli/extract"
+	"github.com/openshift/oc/pkg/cli/logs"
+	"github.com/openshift/oc/pkg/cli/observe"
+	"github.com/openshift/oc/pkg/cli/options"
+	"github.com/openshift/oc/pkg/cli/secrets"
+	"github.com/openshift/oc/pkg/cli/serviceaccounts"
+	"github.com/openshift/oc/pkg/cli/whoami"
 	"github.com/openshift/oc/pkg/helpers/term"
 	"github.com/openshift/origin/pkg/cmd/flagtypes"
 	"github.com/openshift/origin/pkg/cmd/infra/deployer"
@@ -30,20 +39,14 @@ import (
 	"github.com/openshift/origin/pkg/oc/cli/buildlogs"
 	"github.com/openshift/origin/pkg/oc/cli/cancelbuild"
 	"github.com/openshift/origin/pkg/oc/cli/debug"
-	"github.com/openshift/origin/pkg/oc/cli/experimental/dockergc"
-	"github.com/openshift/origin/pkg/oc/cli/expose"
-	"github.com/openshift/origin/pkg/oc/cli/extract"
 	"github.com/openshift/origin/pkg/oc/cli/idle"
 	"github.com/openshift/origin/pkg/oc/cli/image"
 	"github.com/openshift/origin/pkg/oc/cli/importimage"
 	"github.com/openshift/origin/pkg/oc/cli/kubectlwrappers"
 	"github.com/openshift/origin/pkg/oc/cli/login"
 	"github.com/openshift/origin/pkg/oc/cli/logout"
-	"github.com/openshift/origin/pkg/oc/cli/logs"
 	"github.com/openshift/origin/pkg/oc/cli/newapp"
 	"github.com/openshift/origin/pkg/oc/cli/newbuild"
-	"github.com/openshift/origin/pkg/oc/cli/observe"
-	"github.com/openshift/origin/pkg/oc/cli/options"
 	"github.com/openshift/origin/pkg/oc/cli/policy"
 	"github.com/openshift/origin/pkg/oc/cli/process"
 	"github.com/openshift/origin/pkg/oc/cli/project"
@@ -54,13 +57,10 @@ import (
 	"github.com/openshift/origin/pkg/oc/cli/rollout"
 	"github.com/openshift/origin/pkg/oc/cli/rsh"
 	"github.com/openshift/origin/pkg/oc/cli/rsync"
-	"github.com/openshift/origin/pkg/oc/cli/secrets"
-	"github.com/openshift/origin/pkg/oc/cli/serviceaccounts"
 	"github.com/openshift/origin/pkg/oc/cli/set"
 	"github.com/openshift/origin/pkg/oc/cli/startbuild"
 	"github.com/openshift/origin/pkg/oc/cli/status"
 	"github.com/openshift/origin/pkg/oc/cli/tag"
-	"github.com/openshift/origin/pkg/oc/cli/whoami"
 )
 
 const productName = `OpenShift`
