+++ pkg/oc/cli/rsh/rsh.go
@@ -27,8 +27,8 @@ import (
 
 	oapps "github.com/openshift/api/apps"
 	appsv1client "github.com/openshift/client-go/apps/clientset/versioned/typed/apps/v1"
+	cmdutil "github.com/openshift/oc/pkg/helpers/cmd"
 	appsutil "github.com/openshift/origin/pkg/apps/util"
-	"github.com/openshift/origin/pkg/cmd/util"
 )
 
 const (
@@ -203,7 +203,7 @@ func podForResource(f kcmdutil.Factory, resource string, timeout time.Duration) 
 	if err != nil {
 		return "", err
 	}
-	resourceType, name, err := util.ResolveResource(corev1.Resource("pods"), resource, mapper)
+	resourceType, name, err := cmdutil.ResolveResource(corev1.Resource("pods"), resource, mapper)
 	if err != nil {
 		return "", err
 	}
