+++ pkg/oc/cli/set/deploymenthook.go
@@ -19,8 +19,8 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
 
 	appsv1 "github.com/openshift/api/apps/v1"
+	cmdutil "github.com/openshift/oc/pkg/helpers/cmd"
 	utilenv "github.com/openshift/oc/pkg/helpers/env"
-	cmdutil "github.com/openshift/origin/pkg/cmd/util"
 )
 
 var (
