+++ staging/src/github.com/openshift/oc/pkg/cli/create/routepassthrough.go
@@ -10,7 +10,7 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
 
 	routev1 "github.com/openshift/api/route/v1"
-	"github.com/openshift/origin/pkg/oc/cli/create/route"
+	"github.com/openshift/oc/pkg/cli/create/route"
 )
 
 var (
