+++ staging/src/github.com/openshift/oc/pkg/helpers/cmd/sibling.go
@@ -1,4 +1,4 @@
-package util
+package cmd
 
 import (
 	"os"
